Extract TemplateChange class into its own package
Just so it doesn’t get lost amongst all the other template stuff.
