+++ notifications_utils/template.py
@@ -3,6 +3,7 @@ import math
 from notifications_utils.columns import Columns
 from notifications_utils.renderers import SMSPreview, EmailPreview, LetterPreview
 from notifications_utils.field import Field
+from notifications_utils.template_change import TemplateChange
 
 
 class Template():
@@ -134,31 +135,6 @@ class NeededByTemplateError(Exception):
 class NoPlaceholderForDataError(Exception):
     def __init__(self, keys):
         super(NoPlaceholderForDataError, self).__init__(", ".join(keys))
-
-
-class TemplateChange():
-
-    def __init__(self, old_template, new_template):
-        self.old_placeholders = Columns.from_keys(old_template.placeholders)
-        self.new_placeholders = Columns.from_keys(new_template.placeholders)
-
-    @property
-    def has_different_placeholders(self):
-        return bool(self.new_placeholders.keys() ^ self.old_placeholders.keys())
-
-    @property
-    def placeholders_added(self):
-        return set(
-            self.new_placeholders.get(key)
-            for key in self.new_placeholders.keys() - self.old_placeholders.keys()
-        )
-
-    @property
-    def placeholders_removed(self):
-        return set(
-            self.old_placeholders.get(key)
-            for key in self.old_placeholders.keys() - self.new_placeholders.keys()
-        )
 
 
 def get_sms_fragment_count(character_count):
